changed some id fields type from Integer to Long

+++ fakturogen-invoice/src/main/java/pl/fakturogen/invoice/web/dto/InvoiceSaveDTO.java
@@ -27,5 +27,5 @@ public class InvoiceSaveDTO {
     private Integer bankAccountId;
     private Integer invoiceType;
     private String additionalInformation;
-    private Integer originalId; // Document id from provider database
+    private Long originalId; // Document id from provider database
 }
